refactor: move mask sdk
