+++ packages/mask/background/services/wallet/services/helpers.ts
@@ -2,7 +2,7 @@ import type { LegacyWalletRecord } from '@masknet/shared-base'
 import { formatEthereumAddress } from '@masknet/web3-shared-evm'
 import type { WalletDB } from '../database/Wallet.db.js'
 import type { LegacyWalletRecordInDatabase } from '../database/types.js'
-import type { IDBPSafeTransaction } from '../../../../background/database/utils/openDB.js'
+import type { IDBPSafeTransaction } from '../../../database/utils/openDB.js'
 
 export async function getWalletByAddress(t: IDBPSafeTransaction<WalletDB, ['Wallet']>, address: string) {
     const record = await t.objectStore('Wallet').get(formatEthereumAddress(address))
