+++ packages/mask/background/services/wallet/services/legacyWallet.ts
@@ -4,7 +4,7 @@ import { BigNumber } from 'bignumber.js'
 import { ec as EC } from 'elliptic'
 import { fromHex, toHex, type LegacyWalletRecord } from '@masknet/shared-base'
 import { HD_PATH_WITHOUT_INDEX_ETHEREUM } from '@masknet/web3-shared-base'
-import { createTransaction } from '../../../../background/database/utils/openDB.js'
+import { createTransaction } from '../../../database/utils/openDB.js'
 import { createWalletDBAccess } from '../database/Wallet.db.js'
 import { LegacyWalletRecordOutDB } from './helpers.js'
